% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/biomon.R
\name{biomon}
\alias{biomon}
\title{Function to call Biomon's interface}
\usage{
biomon(parm = TRUE)
}
\arguments{
\item{parm}{Parameter just for atomatic test purposes, there is no need to inform it.}
}
\description{
Graphical User Interface for Biomon
}
\examples{
biomon(FALSE)
}

